requirements.txt
alembic==1.4.3
boto3==1.16.13
botocore==1.19.13
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.11
Werkzeug==1.0.1
WTForms==2.3.3
